Create README_Netflix
Objectives
The goal of this project is to develop a recommender system through collaborative filtering. Four different recommender algorithms are used in this study and their performances are evaluated and compared to determine which algorithm is best for Netflix’s business model.

This project uses the Netflix Movies and TV Shows dataset. It is collected from a third-party search engine called Flixable. The dataset consists of the main movie titles table and 4 text files each containing 100,480 ratings from 480,189 customers. The ratings are given to 17,770 movie titles. The text files contain 3 columns consisting of a unique MovieID, customer rating (1 to 5), and date of grade. The movie title dataset consists of 3 columns of the corresponding unique MovieID, year of release and title. Due to limited processing space and speed, only one of the text files are joined to the movie title data through the common MovieID element.

Descriptive statistics and visuals are run on the dataset. By observing the histogram of ratings below, it appears that the data is unbalanced and skewed left.  
A boxplot for ratings shows outliers for movies that rate low or have a score of 1.
 

To evaluate the goodness of fit each of the four algorithms have on the dataset, Root Mean Square Error (RMSE) and Mean Absolute Error (MAE) are measured and compared. RMSE shows the square root of the average squared difference between the predicted and actual values of the movie ratings. MAE shows the mean absolute difference between the predicted and actual value. The lower the RMSE and the MAE, the better the fit . These values provide guidance on the which algorithm is best for Netflix